fix saveModal component and unit tests
